% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.r
\name{dim_reduction}
\alias{dim_reduction}
\title{dimentionality reduction}
\usage{
dim_reduction(
  exprsdata,
  meta.data,
  target_columns = c(2, 3),
  reduction = "PCA",
  save.fig = TRUE,
  output_dir = getwd(),
  pointsize = 3
)
}
\arguments{
\item{exprsdata}{normalized log counts}

\item{meta.data}{meta data for experiment}

\item{target_columns}{meta data to display by shape and color (2 columns can be shown)}

\item{reduction}{what feature reduction algorithm to use PCA (default) or UMAP}

\item{save.fig}{whether to save figures in a file (defualt is TRUE)}

\item{output_dir}{where to save results (default is current working directory)}

\item{pointsize}{size of dots that represent samples on pca or umap (default is 3)}
}
\description{
This implements pca or umap dimensionality reduction on data from normalization step
}
\examples{
data <- hk_gene_stats(counts)


}
\keyword{genes}
\keyword{house}
\keyword{keeping}
