% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_folder.r
\name{generate_folder}
\alias{generate_folder}
\title{generates folder to put results}
\usage{
generate_folder(foldername)
}
\arguments{
\item{foldername}{name of folder (directory) to generate}
}
\description{
generates folder to put results
}
\examples{
generate_folder("directoryname")


}
