% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationorm.r
\name{ratio_normalization}
\alias{ratio_normalization}
\title{ratio normalization}
\usage{
ratio_normalization(
  data,
  hkgenes,
  output_dir = getwd(),
  save.fig = TRUE,
  separte_norm_counts_by_condition = NULL
)
}
\arguments{
\item{hkgenes}{list of hk genes to use to get ratio}

\item{output_dir}{directory for results (default is working directory)}

\item{counts}{matrix counts}
}
\description{
This implements ratio normalization
}
\examples{
data <- hk_gene_stats(counts)
}
\keyword{genes}
\keyword{house}
\keyword{keeping}
