% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialgeneexpression.r
\name{run_DE_analysis}
\alias{run_DE_analysis}
\title{differential gene expression}
\usage{
run_DE_analysis(
  exprsdata,
  meta.data,
  compare.column,
  contrastslist = NULL,
  adj.pval.method = "BH",
  adjbycomparison = TRUE,
  pval.cutoff = 0.05,
  save.fig = TRUE,
  check.norm = TRUE,
  DE.test = "ANOVA",
  covariate.column = NULL,
  output_dir = getwd(),
  volcano.plot = TRUE,
  heatmap.plot = TRUE
)
}
\arguments{
\item{exprsdata}{normalized log counts}

\item{meta.data}{meta data for experiment}

\item{compare.column}{column from meta data you will be comparing (i.e. treatmennt differences the user wants to explore)}

\item{contrastslist}{list of contrasts of treatments (i.e. c(t"reatment1-control", "treatment2-control"))}

\item{adj.pval.method}{method used to adjust pvalues for mulitiple hypothesis (default is BH)}

\item{adjbycomparison}{whether to adjust pvalues by each comparison seperately (default is this) or all together for ttest or wilcox}

\item{pval.cutoff}{pvalue cutoff for determining sigificant genes (default is 0.05)}

\item{save.fig}{whether to save figures in PDF and PNG file format}

\item{check.norm}{whether to check if data follows  a normal distribution (default is TRUE)}

\item{DE.test}{which significance test to use (ttest, wilcox or ANOVA/tukeys(default))}

\item{covariate.column}{column in meta data to use as covariate (ONLY CAN BE USED IN ANOVA TEST)}

\item{output_dir}{directory to store results}

\item{volcano.plot}{whether to make volcano plots showing DE genes (default is TRUE)}

\item{heatmap.plot}{whether to make heatmap showing DE genes (default is TRUE)}
}
\description{
Implements a t.test or annova tukey's test to determine DE genes
}
