% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationorm.r
\name{ratio_normalization}
\alias{ratio_normalization}
\title{ratio normalization}
\usage{
ratio_normalization(data, hkgenes, output_dir = getwd())
}
\arguments{
\item{counts}{matrix counts}
}
\description{
This implements ratio normalization
}
\examples{
data <- hk_gene_stats(counts)
}
\keyword{genes}
\keyword{house}
\keyword{keeping}
